refactor: changed login and register schemas names to loginSchema and registerSchema, to avoid conflict with react-hook-form
