+++ src/shared/schemas/index.ts
@@ -1,2 +1,2 @@
-export * from "./login";
-export * from "./register";
+export * from "./loginSchema";
+export * from "./registerSchema";
