+++ src/shared/schemas/registerSchema.ts
